requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.0
altair==4.2.2
anyio==3.6.2
apache-airflow==2.5.1
apache-airflow-providers-amazon==8.1.0
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.3.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.2
apache-airflow-providers-tableau==4.3.0
apispec==3.3.2
argcomplete==3.0.8
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.6.2
boto3==1.26.155
botocore==1.29.165
cachelib==0.9.0
cattrs==22.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.2.2
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.3.0
croniter==1.3.14
cryptography==40.0.2
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
dnspython==2.3.0
docutils==0.20.1
email-validator==1.3.1
entrypoints==0.4
exceptiongroup==1.1.1
executing==2.0.1
fastjsonschema==2.20.0
Flask==2.2.5
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.10.0
graphviz==0.20.1

greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflection==0.5.1
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.25
jsonpatch==1.33
jsonpath-ng==1.5.3
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
lazy-object-proxy==1.9.0
limits==3.5.0
linkify-it-py==2.0.2
lockfile==0.12.2
lxml==4.9.2
makefun==1.15.3
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.5
mdurl==0.1.2
mistune==3.0.2
MouseInfo==0.1.3
multidict==6.0.4
mypy-boto3-appflow==1.26.125
mypy-boto3-rds==1.26.136
mypy-boto3-redshift-data==1.26.109
mypy-boto3-s3==1.26.127
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
ordered-set==4.1.0
overrides==7.7.0
packaging==23.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.0.0
ply==3.11
prison==0.2.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==5.9.5
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyathena==2.23.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==1.10.7
PyGetWindow==0.0.9
Pygments==2.15.1
PyJWT==2.7.0
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.9.0
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.30
pyspark==3.5.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-nvd3==0.15.0
python-slugify==8.0.1
python3-xlib==0.15
pytweening==1.2.0
pytz==2023.3
pytzdata==2020.1
pywhatkit==5.4
PyYAML==6.0
pyzmq==26.0.3
redshift-connector==2.0.910
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rich_argparse==1.1.0
rpds-py==0.19.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.8
s3transfer==0.6.1
scipy==1.13.1
scramp==1.4.4
Send2Trash==1.8.3
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.48
SQLAlchemy-JSONField==1.0.1.post0
sqlalchemy-redshift==0.8.14
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.3
tableauserverclient==0.25
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.3.0
tomli==2.0.1
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.5.0
tzdata==2024.1
tzlocal==5.2
uc-micro-py==1.0.2
unicodecsv==0.14.1
uri-template==1.3.0
urllib3==1.26.16
watchtower==2.0.1
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.3
widgetsnbextension==4.0.11
wikipedia==1.4.0
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
ytmusicapi==1.7.3
zipp==3.15.0
# Kafka dependencies
kafka-python==2.0.2
confluent-kafka==2.3.0
